+++ games/net.c
@@ -23,9 +23,6 @@
  * works better on stylus-driven platforms such as Palm and
  * PocketPC, though, so we enable it by default there.
  */
-#ifdef STYLUS_BASED
-#define USE_DRAGGING
-#endif
 
 /* Direction and other bitfields */
 #define R 0x01
@@ -1597,11 +1594,14 @@ static const char *validate_desc(const game_params *params, const char *desc)
 
 static key_label *game_request_keys(const game_params *params, int *nkeys)
 {
-    key_label *keys = snewn(1, key_label);
-    *nkeys = 1;
+    key_label *keys = snewn(2, key_label);
+    *nkeys = 2;
 
     keys[0].button = 'J';
     keys[0].label = "Shuffle";
+
+    keys[1].button = 'L';
+    keys[1].label = "Lock";
 
     return keys;
 }
@@ -1979,13 +1979,8 @@ static int *compute_loops(const game_state *state)
 struct game_ui {
     int org_x, org_y; /* origin */
     int cx, cy;       /* source tile (game coordinates) */
-    int cur_x, cur_y;
-    bool cur_visible;
     random_state *rs; /* used for jumbling */
-#ifdef USE_DRAGGING
-    int dragtilex, dragtiley, dragstartx, dragstarty;
-    bool dragged;
-#endif
+    bool use_locking;
 };
 
 static game_ui *new_ui(const game_state *state)
@@ -1994,9 +1989,9 @@ static game_ui *new_ui(const game_state *state)
     int seedsize;
     game_ui *ui = snew(game_ui);
     ui->org_x = ui->org_y = 0;
-    ui->cur_x = ui->cx = state->width / 2;
-    ui->cur_y = ui->cy = state->height / 2;
-    ui->cur_visible = false;
+    ui->cx = state->width / 2;
+    ui->cy = state->height / 2;
+    ui->use_locking = false;
     get_random_seed(&seed, &seedsize);
     ui->rs = random_new(seed, seedsize);
     sfree(seed);
@@ -2021,15 +2016,18 @@ static char *encode_ui(const game_ui *ui)
     return dupstr(buf);
 }
 
-static void decode_ui(game_ui *ui, const char *encoding)
-{
+static void decode_ui(game_ui *ui, const char *encoding) {
     sscanf(encoding, "O%d,%d;C%d,%d",
        &ui->org_x, &ui->org_y, &ui->cx, &ui->cy);
 }
 
 static void game_changed_state(game_ui *ui, const game_state *oldstate,
-                               const game_state *newstate)
-{
+                               const game_state *newstate) {
+}
+
+static bool is_key_highlighted(const game_ui *ui, char c) {
+    if (c == 'L') return ui->use_locking;
+    return false;
 }
 
 struct game_drawstate {
@@ -2046,158 +2044,44 @@ static char *interpret_move(const game_state *state, game_ui *ui,
                             int x, int y, int button, bool swapped)
 {
     char *nullret;
-    int tx = -1, ty = -1, dir = 0;
+    int tx = -1, ty = -1;
     enum {
-        NONE, ROTATE_LEFT, ROTATE_180, ROTATE_RIGHT, TOGGLE_LOCK, JUMBLE,
-        MOVE_ORIGIN, MOVE_SOURCE, MOVE_ORIGIN_AND_SOURCE, MOVE_CURSOR
+        NONE, ROTATE_LEFT, ROTATE_RIGHT, TOGGLE_LOCK, JUMBLE
     } action;
 
     button &= ~MOD_MASK;
     nullret = NULL;
     action = NONE;
 
-    if (button == LEFT_BUTTON ||
-    button == MIDDLE_BUTTON ||
-#ifdef USE_DRAGGING
-    button == LEFT_DRAG ||
-    button == LEFT_RELEASE ||
-    button == RIGHT_DRAG ||
-    button == RIGHT_RELEASE ||
-#endif
-    button == RIGHT_BUTTON) {
+    if (button == LEFT_BUTTON || button == RIGHT_BUTTON) {
+        /*
+         * The button must have been clicked on a valid tile.
+         */
+        x -= WINDOW_OFFSET + LINE_THICK;
+        y -= WINDOW_OFFSET + LINE_THICK;
+        if (x < 0 || y < 0)
+            return nullret;
+        tx = x / TILE_SIZE;
+        ty = y / TILE_SIZE;
+        if (tx >= state->width || ty >= state->height)
+            return nullret;
+        /* Transform from physical to game coords */
+        tx = (tx + ui->org_x) % state->width;
+        ty = (ty + ui->org_y) % state->height;
+        if (x % TILE_SIZE >= TILE_SIZE - LINE_THICK ||
+            y % TILE_SIZE >= TILE_SIZE - LINE_THICK)
+            return nullret;
 
-    if (ui->cur_visible) {
-        ui->cur_visible = false;
-        nullret = UI_UPDATE;
-    }
+        if (ui->use_locking)
+            action = (button == LEFT_BUTTON ? ROTATE_LEFT : TOGGLE_LOCK);
+        else
+            action = (button == LEFT_BUTTON ? ROTATE_LEFT : ROTATE_RIGHT);
 
-    /*
-     * The button must have been clicked on a valid tile.
-     */
-    x -= WINDOW_OFFSET + LINE_THICK;
-    y -= WINDOW_OFFSET + LINE_THICK;
-    if (x < 0 || y < 0)
-        return nullret;
-    tx = x / TILE_SIZE;
-    ty = y / TILE_SIZE;
-    if (tx >= state->width || ty >= state->height)
-        return nullret;
-    /* Transform from physical to game coords */
-    tx = (tx + ui->org_x) % state->width;
-    ty = (ty + ui->org_y) % state->height;
-    if (x % TILE_SIZE >= TILE_SIZE - LINE_THICK ||
-        y % TILE_SIZE >= TILE_SIZE - LINE_THICK)
-        return nullret;
-
-#ifdef USE_DRAGGING
-
-        if (button == MIDDLE_BUTTON
-#ifdef STYLUS_BASED
-        || button == RIGHT_BUTTON  /* with a stylus, `right-click' locks */
-#endif
-        ) {
-            /*
-             * Middle button never drags: it only toggles the lock.
-             */
-            action = TOGGLE_LOCK;
-        } else if (button == LEFT_BUTTON
-#ifndef STYLUS_BASED
-                   || button == RIGHT_BUTTON /* (see above) */
-#endif
-                  ) {
-            /*
-             * Otherwise, we note down the start point for a drag.
-             */
-            ui->dragtilex = tx;
-            ui->dragtiley = ty;
-            ui->dragstartx = x % TILE_SIZE;
-            ui->dragstarty = y % TILE_SIZE;
-            ui->dragged = false;
-            return nullret;            /* no actual action */
-        } else if (button == LEFT_DRAG
-#ifndef STYLUS_BASED
-                   || button == RIGHT_DRAG
-#endif
-                  ) {
-            /*
-             * Find the new drag point and see if it necessitates a
-             * rotation.
-             */
-            int x0,y0, xA,yA, xC,yC, xF,yF;
-            int mx, my;
-            int d0, dA, dC, dF, dmin;
-
-            tx = ui->dragtilex;
-            ty = ui->dragtiley;
-
-            mx = x - (ui->dragtilex * TILE_SIZE);
-            my = y - (ui->dragtiley * TILE_SIZE);
-
-            x0 = ui->dragstartx;
-            y0 = ui->dragstarty;
-            xA = ui->dragstarty;
-            yA = TILE_SIZE-1 - ui->dragstartx;
-            xF = TILE_SIZE-1 - ui->dragstartx;
-            yF = TILE_SIZE-1 - ui->dragstarty;
-            xC = TILE_SIZE-1 - ui->dragstarty;
-            yC = ui->dragstartx;
-
-            d0 = (mx-x0)*(mx-x0) + (my-y0)*(my-y0);
-            dA = (mx-xA)*(mx-xA) + (my-yA)*(my-yA);
-            dF = (mx-xF)*(mx-xF) + (my-yF)*(my-yF);
-            dC = (mx-xC)*(mx-xC) + (my-yC)*(my-yC);
-
-            dmin = min(min(d0,dA),min(dF,dC));
-
-            if (d0 == dmin) {
-                return nullret;
-            } else if (dF == dmin) {
-                action = ROTATE_180;
-                ui->dragstartx = xF;
-                ui->dragstarty = yF;
-                ui->dragged = true;
-            } else if (dA == dmin) {
-                action = ROTATE_LEFT;
-                ui->dragstartx = xA;
-                ui->dragstarty = yA;
-                ui->dragged = true;
-            } else /* dC == dmin */ {
-                action = ROTATE_RIGHT;
-                ui->dragstartx = xC;
-                ui->dragstarty = yC;
-                ui->dragged = true;
-            }
-        } else if (button == LEFT_RELEASE
-#ifndef STYLUS_BASED
-                   || button == RIGHT_RELEASE
-#endif
-                  ) {
-            if (!ui->dragged) {
-                /*
-                 * There was a click but no perceptible drag:
-                 * revert to single-click behaviour.
-                 */
-                tx = ui->dragtilex;
-                ty = ui->dragtiley;
-
-                if (button == LEFT_RELEASE)
-                    action = ROTATE_LEFT;
-                else
-                    action = ROTATE_RIGHT;
-            } else
-                return nullret;        /* no action */
-        }
-
-#else /* USE_DRAGGING */
-
-    action = (button == LEFT_BUTTON ? ROTATE_LEFT :
-          button == RIGHT_BUTTON ? ROTATE_RIGHT : TOGGLE_LOCK);
-
-#endif /* USE_DRAGGING */
-
-    } else if (button == 'j' || button == 'J') {
-        /* XXX should we have some mouse control for this? */
+    } else if (button == 'J') {
         action = JUMBLE;
+    } else if (button == 'L') {
+        ui->use_locking = !ui->use_locking;
+        return UI_UPDATE;
     } else
         return nullret;
 
@@ -2215,8 +2099,7 @@ static char *interpret_move(const game_state *state, game_ui *ui,
         char buf[80];
         sprintf(buf, "L%d,%d", tx, ty);
         return dupstr(buf);
-    } else if (action == ROTATE_LEFT || action == ROTATE_RIGHT ||
-               action == ROTATE_180) {
+    } else if (action == ROTATE_LEFT || action == ROTATE_RIGHT) {
         char buf[80];
 
         /*
@@ -2266,22 +2149,6 @@ static char *interpret_move(const game_state *state, game_ui *ui,
         ret = sresize(ret, p - ret, char);
 
         return ret;
-    } else if (action == MOVE_ORIGIN || action == MOVE_SOURCE ||
-               action == MOVE_ORIGIN_AND_SOURCE || action == MOVE_CURSOR) {
-        assert(dir != 0);
-        if (action == MOVE_ORIGIN || action == MOVE_ORIGIN_AND_SOURCE) {
-            if (state->wrapping) {
-                 OFFSET(ui->org_x, ui->org_y, ui->org_x, ui->org_y, dir, state);
-            } else return nullret; /* disallowed for non-wrapping grids */
-        }
-        if (action == MOVE_SOURCE || action == MOVE_ORIGIN_AND_SOURCE) {
-            OFFSET(ui->cx, ui->cy, ui->cx, ui->cy, dir, state);
-        }
-        if (action == MOVE_CURSOR) {
-            OFFSET(ui->cur_x, ui->cur_y, ui->cur_x, ui->cur_y, dir, state);
-            ui->cur_visible = true;
-        }
-        return UI_UPDATE;
     } else {
         return NULL;
     }
@@ -2407,6 +2274,7 @@ static game_drawstate *game_new_drawstate(drawing *dr, const game_state *state)
 static void game_free_drawstate(drawing *dr, game_drawstate *ds)
 {
     sfree(ds->visible);
+    sfree(ds->to_draw);
     sfree(ds);
 }
 
@@ -2462,7 +2330,6 @@ static void rotated_coords(float *ox, float *oy, const float matrix[4],
 /* Flags describing the visible features of a tile. */
 #define TILE_BARRIER_SHIFT            0  /* 4 bits: R U L D */
 #define TILE_BARRIER_CORNER_SHIFT     4  /* 4 bits: RU UL LD DR */
-#define TILE_KEYBOARD_CURSOR      (1<<8) /* 1 bit if cursor is here */
 #define TILE_WIRE_SHIFT               9  /* 8 bits: RR UU LL DD
                                           * Each pair: 0=no wire, 1=unpowered,
                                           * 2=powered, 3=error highlight */
@@ -2820,9 +2687,6 @@ static void game_redraw(drawing *dr, game_drawstate *ds,
                 }
             }
 
-            if (ui->cur_visible && gx == ui->cur_x && gy == ui->cur_y)
-                todraw(ds, dx, dy) |= TILE_KEYBOARD_CURSOR;
-
             if (gx == tx && gy == ty)
                 todraw(ds, dx, dy) |= TILE_ROTATING;
 
@@ -2952,8 +2816,10 @@ static bool game_timing_state(const game_state *state, game_ui *ui)
 #define thegame net
 #endif
 
-static const char rules[] = "The computer prepares a network by connecting up the centres of squares in a grid, and then shuffles the network by rotating every tile randomly.\n\n"
-"Your job is to rotate it all back into place. The successful solution will be an entirely connected network, with no closed loops. As a visual aid, all tiles which are connected to the one in the middle are highlighted.\n\n\n"
+static const char rules[] = "A network is prepared by connecting up the centres of squares in a grid, and then shuffled by rotating every tile randomly.\n\n"
+"The task is to rotate it all back into place. The successful solution will be an entirely connected network, with no closed loops. As a visual aid, all tiles which are connected to the one in the middle are highlighted.\n\n"
+"Lock button: When this button is active, you can lock a tile in place with a long-click. Long-click again unlocks it.\n\n"
+"Jumble button: This button turns all tiles that are not locked to random orientations.\n\n\n"
 "This puzzle was implemented by Simon Tatham.";
 
 const struct game thegame = {
@@ -2989,7 +2855,7 @@ const struct game thegame = {
     game_anim_length,
     game_flash_length,
     NULL,
-    NULL,
+    is_key_highlighted,
     game_status,
     false, false, NULL, NULL,
     true,                   /* wants_statusbar */