+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 ### Changed
 * All games: Use status bar to show game completion instead of pushy message box
+* *Net*: Add lock button for locking tiles in place where one is sure of the orientation.
 
 ## 0.7.10 - 2022/04/02